抽取文本编辑组件
