+++ index.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <div id="app" class="w-[100vw] h-[100vh]"></div>
-    <script type="module" src="/src/main.ts"></script>
-    <!-- <script type="module" src="/demo/src/main.ts"></script> -->
+    <!-- <script type="module" src="/src/main.ts"></script> -->
+    <script type="module" src="/demo/src/main.ts"></script>
   </body>
 </html>
